feat: adding new categories and note titles
It is possible to add categories and notes (only titles, not the note itself at this point). Nothing is stored anywhere, not even in local storage. So whenever you reload the page, you will start without any categories or notes at this point. Next step would be to implement a feature that lets you add note texts via markdown.